mv get_content to utils
