+++ crates/gdtk/src/commands/dev/lex.rs
@@ -2,7 +2,7 @@ use std::path::PathBuf;
 
 use itertools::Itertools;
 
-use crate::commands::get_content;
+use crate::utils::get_content;
 
 pub fn run(file: PathBuf) -> anyhow::Result<()> {
     let content = get_content(file.as_path())?;
