+++ crates/gdtk/src/commands/dev/parse.rs
@@ -1,6 +1,6 @@
 use std::path::PathBuf;
 
-use crate::commands::get_content;
+use crate::utils::get_content;
 
 pub fn run(file: PathBuf) -> anyhow::Result<()> {
     let content = get_content(file.as_path())?;
