+++ crates/gdtk/src/lib.rs
@@ -3,3 +3,4 @@
 pub mod cli;
 pub mod commands;
 pub mod display;
+pub mod utils;
